Create Mini Browser
UM BROWSER is an android application for accessing information on the World Wide Web. When a user requests a web page from a particular website, the UM BROWSER retrieves the necessary content from a web server and then displays the page on the user's device. This process begins when the user inputs a Uniform Resource Locator (URL), such as https://www.google.com/, into the browser. And once a web page has been retrieved, the browser's rendering engine displays it on the user's device. This includes image and video formats supported by the browser. UM BROWSER is supported on android smartphone devices ranging from android 5.0 to latest.